Organize file structure divided by components & containers
