+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import App from './App'
+import App from './containers/App'
 import reportWebVitals from "./reportWebVitals";
 import "tachyons";
